Import jstz.min.js from jstimezonedetect 1.0.7
In Nextcloud 22 jstimezonedetect was removed from the server bundle
(https://github.com/nextcloud/server/pull/25850), so now each app that
uses it needs to ship its own copy. For simplicity, as the app does not
currently use webpack or NPM, the minified file was just added to
"js/vendor".

"jstz.min.js" was copied from
https://github.com/pellepim/jstimezonedetect/blob/ddc9e040342d68c06129ae28011517e5b2dde664/dist/jstz.min.js
(which is the same as the one included in
https://registry.npmjs.org/jstimezonedetect/-/jstimezonedetect-1.0.7.tgz).

jstimezonedetect is licenced under the MIT licence.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>